refactor(pkg): move integrations to internal/
- Move pkg/integrations/ → internal/integrations/
- Update all import paths to maintain functionality
- Follows Level 3 → internal strategy from design
- Build verification passes

Part 1/4 of pkg/ restructure for architecture compliance
