+++ docs/ebpf-audit-report.md
@@ -1,0 +1,268 @@
+# eBPF Implementation Audit Report - Tapio Observability Platform
+
+## Executive Summary
+Audit of eBPF implementations across all Tapio observers for CO-RE compliance, safety, completeness, and best practices.
+
+## 1. CO-RE (Compile Once Run Everywhere) Compliance ❌
+
+### Current Status: **NOT CO-RE COMPLIANT**
+
+**Critical Issues:**
+- Missing `-target bpf` flag in bpf2go commands (required for CO-RE)
+- Not using BTF (BPF Type Format) for portability
+- Using kernel headers directly instead of BTF-based types
+- No `vmlinux.h` generation from BTF
+
+**Evidence:**
+```bash
+# Current bpf2go commands missing CO-RE flags:
+-target amd64,arm64  # Architecture-specific, not CO-RE
+# Should be:
+-target bpf  # For CO-RE compliance
+```
+
+### Observers Using Some CO-RE Features:
+- **systemd**: Uses `BPF_CORE_READ()` macros (partial CO-RE)
+- **health**: Uses `BPF_CORE_READ()` macros (partial CO-RE)
+- **memory**: Includes `bpf_core_read.h` but doesn't use CO-RE macros
+- **container-runtime**: Includes headers but no CO-RE usage
+
+### Observers NOT Using CO-RE:
+- **dns**: No CO-RE macros or BTF usage
+- **kernel**: No CO-RE, using direct kernel struct access
+- **network**: No CO-RE implementation
+- **process-signals**: No CO-RE
+- **storage-io**: No CO-RE
+- **scheduler**: No CO-RE
+
+## 2. Implementation Completeness 🟡
+
+### Fully Implemented Observers:
+1. **DNS Observer** ✅
+   - Complete DNS packet parsing
+   - IPv4/IPv6 support
+   - UDP/TCP support
+   - Query/Response correlation
+   - Multiple DNS record types
+
+2. **Memory Observer** ✅
+   - Allocation tracking
+   - RSS monitoring
+   - Stack trace collection
+   - Memory leak detection patterns
+
+3. **Container Runtime Observer** ✅
+   - OOM kill detection
+   - Process lifecycle tracking
+   - Container ID extraction
+   - Cgroup tracking
+
+### Partially Implemented:
+1. **Kernel Observer** 🟡
+   - Only ConfigMap/Secret access tracking
+   - Missing general syscall monitoring
+   - Limited to specific mount patterns
+
+2. **Network Observer** 🟡
+   - Basic TCP/UDP tracking
+   - Missing L7 protocol parsing
+   - No SSL/TLS visibility
+
+3. **Storage I/O Observer** 🟡
+   - Multiple versions (simple, premium)
+   - Inconsistent implementation across versions
+
+### Stub/Incomplete Implementations:
+1. **Scheduler Observer** ❌ - BPF source exists but no eBPF loader
+2. **Services Observer** ❌ - Has ebpf.go but no BPF source
+3. **Status Observer** ❌ - Minimal stub implementation
+
+## 3. Missing Best Practices (MVP Considerations)
+
+### Critical Missing Elements:
+
+#### A. Error Handling & Safety ⚠️
+```c
+// MISSING: Bounds checking before string reads
+bpf_probe_read_user_str(buf, sizeof(buf), path);  // No null check
+
+// MISSING: Return value checks
+bpf_perf_event_output(ctx, &events, ...);  // Ignoring failures
+```
+
+#### B. Resource Management ⚠️
+- No consistent ring buffer size limits
+- Some maps too large: `1024 * 1024` entries
+- Missing memory pressure handling
+- No rate limiting in most observers
+
+#### C. Verifier Compliance 🟡
+- Missing loop bounds in some cases
+- Complex nested conditions that may hit verifier limits
+- No `#pragma unroll` for loops where needed
+
+#### D. Performance Optimization ❌
+- Not using per-CPU maps for high-frequency events
+- Missing batch processing in high-volume observers
+- No sampling for expensive operations
+
+## 4. Safety & Standards Issues 🔴
+
+### Critical Safety Issues:
+
+#### A. Memory Safety
+```c
+// UNSAFE: Direct memory access without bounds
+char comm[16];
+bpf_get_current_comm(&comm, sizeof(comm));  // OK
+strcpy(event->comm, comm);  // UNSAFE - should use bpf_probe_read
+```
+
+#### B. Kernel Stability Risks
+- **No BPF_PROG_TYPE verification** in some programs
+- **Missing SEC() annotations** for program types
+- **No capability checks** before privileged operations
+
+#### C. Security Issues
+- **No PID namespace awareness** in most observers
+- **Missing container boundary checks**
+- **No rate limiting** for DoS prevention
+- **Sensitive data exposure** (ConfigMaps, Secrets) without filtering
+
+### Compliance Issues:
+
+#### GPL Licensing ✅
+All BPF programs correctly use `// SPDX-License-Identifier: GPL-2.0`
+
+#### Missing Standard Headers ⚠️
+```c
+// Should include for safety:
+#include <linux/bpf.h>
+#include <linux/ptrace.h>
+#include <linux/sched.h>
+```
+
+## 5. Recommendations for MVP
+
+### Immediate Actions (P0):
+
+1. **Add Safety Checks**
+```c
+// Before any probe read:
+if (!path) return 0;
+int ret = bpf_probe_read_user_str(buf, sizeof(buf), path);
+if (ret < 0) return 0;
+```
+
+2. **Add Ring Buffer Overflow Handling**
+```c
+struct event *e = bpf_ringbuf_reserve(&events, sizeof(*e), 0);
+if (!e) {
+    __sync_fetch_and_add(&dropped_events, 1);
+    return 0;
+}
+```
+
+3. **Implement Rate Limiting**
+```c
+// Per-CPU rate limit map
+struct bpf_map_def SEC("maps") rate_limit = {
+    .type = BPF_MAP_TYPE_PERCPU_ARRAY,
+    .max_entries = 1,
+};
+```
+
+### Short-term (P1):
+
+1. **Migrate to CO-RE**
+   - Generate `vmlinux.h` from BTF
+   - Replace struct access with `BPF_CORE_READ()`
+   - Use `-target bpf` in bpf2go
+
+2. **Add Sampling**
+```c
+// Sample 1 in N events for high-volume observers
+if (bpf_get_prandom_u32() % SAMPLE_RATE != 0)
+    return 0;
+```
+
+3. **Standardize Map Sizes**
+   - Ring buffers: 256KB for normal, 1MB for high-volume
+   - Hash maps: 10K entries max
+   - Per-CPU arrays for counters
+
+### Long-term (P2):
+
+1. **Implement BTF-based CO-RE**
+2. **Add eBPF program chaining**
+3. **Implement tail calls for complex logic**
+4. **Add BPF-to-BPF function calls**
+
+## 6. Per-Observer Action Items
+
+### DNS Observer
+- ✅ Mostly complete
+- Add rate limiting for high QPS
+- Add sampling for performance
+
+### Memory Observer  
+- Add CO-RE support
+- Reduce allocation map size (currently 1M entries)
+- Add PID namespace filtering
+
+### Container Runtime
+- Add CO-RE support
+- Implement proper container boundary checks
+- Add rate limiting for fork/exec events
+
+### Network Observer
+- Complete L7 protocol parsing
+- Add SSL/TLS support via uprobe
+- Implement connection tracking limits
+
+### Kernel Observer
+- Expand beyond ConfigMap/Secret monitoring
+- Add general syscall monitoring
+- Implement audit trail
+
+### Health Observer
+- Complete implementation (currently stub)
+- Add system health metrics
+- Implement anomaly detection
+
+### Systemd Observer
+- Good CO-RE usage
+- Add service dependency tracking
+- Implement journal correlation
+
+### Storage I/O
+- Consolidate multiple versions
+- Add bio layer monitoring
+- Implement I/O pattern detection
+
+### Scheduler Observer
+- Complete eBPF implementation
+- Add CFS monitoring
+- Track scheduling latency
+
+### Process Signals
+- Add signal correlation
+- Implement signal storm detection
+- Add process tree tracking
+
+## Conclusion
+
+The Tapio eBPF implementations are functional but lack critical safety and portability features for production:
+
+1. **Not CO-RE compliant** - Will break across kernel versions
+2. **Missing safety checks** - Risk of crashes/panics
+3. **No rate limiting** - DoS vulnerability
+4. **Incomplete observers** - Several stubs or partial implementations
+
+For MVP, focus on:
+- Adding safety checks (P0)
+- Implementing rate limiting (P0)
+- Completing critical observers (P1)
+- Planning CO-RE migration (P2)
+
+The architecture is solid, but implementation needs hardening before production deployment.
