+++ k8s/observer-config.yaml
@@ -1,0 +1,13 @@
+orchestrator:
+  workers: 4
+  buffer_size: 10000
+  log_level: "info"
+  
+observers:
+  process-signals:
+    enabled: true
+    config:
+      process_filter: ""
+      signal_types: ["SIGKILL", "SIGTERM"]
+      sample_rate: 1.0
+      buffer_size: 1000
