+++ k8s/observer-deployment.yaml
@@ -1,0 +1,73 @@
+apiVersion: v1
+kind: ConfigMap
+metadata:
+  name: observer-config
+data:
+  config.yaml: |
+    orchestrator:
+      workers: 4
+      buffer_size: 10000
+      log_level: "info"
+      
+    observers:
+      process-signals:
+        enabled: true
+        config:
+          process_filter: ""
+          signal_types: ["SIGKILL", "SIGTERM"]
+          sample_rate: 1.0
+          buffer_size: 1000
+---
+apiVersion: apps/v1
+kind: Deployment
+metadata:
+  name: tapio-observer
+  labels:
+    app: tapio-observer
+spec:
+  replicas: 1
+  selector:
+    matchLabels:
+      app: tapio-observer
+  template:
+    metadata:
+      labels:
+        app: tapio-observer
+    spec:
+      containers:
+      - name: observer
+        image: tapio-observer:latest
+        imagePullPolicy: Never
+        command: ["/observers"]
+        args: ["-config", "/config/config.yaml", "-log-level", "info"]
+        volumeMounts:
+        - name: config
+          mountPath: /config
+        resources:
+          requests:
+            memory: "64Mi"
+            cpu: "250m"
+          limits:
+            memory: "128Mi"
+            cpu: "500m"
+        securityContext:
+          runAsNonRoot: false
+          privileged: true  # Needed for eBPF if it tries to load
+      volumes:
+      - name: config
+        configMap:
+          name: observer-config
+      restartPolicy: Always
+---
+apiVersion: v1
+kind: Service
+metadata:
+  name: tapio-observer-service
+spec:
+  selector:
+    app: tapio-observer
+  ports:
+  - port: 8080
+    targetPort: 8080
+    name: http
+  type: ClusterIP
